# controller/alert_silence.go
package controller

import (
	"github.com/alert666/api-server/base/bind"
	v1 "github.com/alert666/api-server/service/v1"
	"github.com/gin-gonic/gin"
)

type AlertSilenceController interface {
	CreateAlertSilence(c *gin.Context)
	DeleteAlertSilence(c *gin.Context)
	QueryAlertSilence(c *gin.Context)
	ListAlertSilence(c *gin.Context)
	GetTenantSilenceCounts(c *gin.Context)
}

type alertSilenceController struct {
	alertSilenceImpl v1.AlertSilenceServicer
}

func NewAlertSilenceController(alertSilenceService v1.AlertSilenceServicer) AlertSilenceController {
	return &alertSilenceController{
		alertSilenceImpl: alertSilenceService,
	}
}

// CreateApi 创建 AlerSilence
// @Summary 创建 AlerSilence
// @Description 创建 AlerSilence
// @Tags AlerSilence 管理
// @Accept json
// @Produce json
// @Param data body types.AlertSilenceCreateRequest true "创建请求参数"
// @Success 200 {object} types.Response "创建成功"
// @Router /api/v1/alertSilence [post]
func (receiver *alertSilenceController) CreateAlertSilence(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.alertSilenceImpl.CreateSilence, bind.BindTypeJson)
}

// DeleteApi 删除 AlerSilence
// @Summary 删除 AlerSilence
// @Description 删除 AlerSilence
// @Tags AlerSilence 管理
// @Accept json
// @Produce json
// @Param id path int true "静默规则ID"
// @Success 200 {object} types.Response "删除成功"
// @Router /api/v1/alertSilence/:id [delete]
func (receiver *alertSilenceController) DeleteAlertSilence(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.alertSilenceImpl.DeleteSilence, bind.BindTypeUri)
}

// QueryApi 查询 AlerSilence
// @Summary 查询 AlerSilence
// @Description 查询 AlerSilence
// @Tags AlerSilence 管理
// @Accept json
// @Produce json
// @Param id path int true "静默规则ID"
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/alertSilence/:id [get]
func (receiver *alertSilenceController) QueryAlertSilence(c *gin.Context) {
	bind.ResponseWithData(c, receiver.alertSilenceImpl.QuerySilence, bind.BindTypeUri)
}

// @Summary 获取所有 AlerSilence
// @Description 获取所有 AlerSilence
// @Tags AlerSilence 管理
// @Accept json
// @Produce json
// @Param data query types.AlertSilenceListRequest true "List请求参数"
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/alertSilence [get]
func (receiver *alertSilenceController) ListAlertSilence(c *gin.Context) {
	bind.ResponseWithData(c, receiver.alertSilenceImpl.ListSilence, bind.BindTypeQuery)
}

// @Summary 获取各个租户活跃 AlerSilence 的数量
// @Description 获取各个租户活跃 AlerSilence 的数量
// @Tags AlerSilence 管理
// @Accept json
// @Produce json
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/alertSilence/enableCount [get]
func (receiver *alertSilenceController) GetTenantSilenceCounts(c *gin.Context) {
	bind.ResponseWithDataNoBind(c, receiver.alertSilenceImpl.GetTenantSilenceCounts)
}