# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "告警管理后端服务 API 文档，支持告警接收、静默、抑制、通知、多租户管理、RBAC 权限控制等功能。",
        "title": "api-server — 告警管理 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "paths": {
...
        "/api/v1/alertSilence": {
            "get": {
                "description": "获取所有 AlerSilence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerSilence 管理"
                ],
                "summary": "获取所有 AlerSilence",
                "parameters": [
                    {
                        "type": "string",
                        "name": "createdBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "endsAt",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page 从 1 开始, 表示第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "PageSize 最大值 100，默认值 20, 表示每页多少条数据",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "startsAt",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建 AlerSilence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerSilence 管理"
                ],
                "summary": "创建 AlerSilence",
                "parameters": [
                    {
                        "description": "创建请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertSilenceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/alertSilence/:id": {
            "get": {
                "description": "查询 AlerSilence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerSilence 管理"
                ],
                "summary": "查询 AlerSilence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "静默规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除 AlerSilence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerSilence 管理"
                ],
                "summary": "删除 AlerSilence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "静默规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/alertTemplate": {
            "get": {
                "description": "获取所有 AlerTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerTemplate 管理"
                ],
                "summary": "获取所有 AlerTemplate",
                "parameters": [
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page 从 1 开始, 表示第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "PageSize 最大值 100，默认值 20, 表示每页多少条数据",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建 AlerTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerTemplate 管理"
                ],
                "summary": "创建 AlerTemplate",
                "parameters": [
                    {
                        "description": "创建请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertTemplateCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/alertTemplate/:id": {
            "get": {
                "description": "查询 AlerTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerTemplate 管理"
                ],
                "summary": "查询 AlerTemplate",
                "parameters": [
                    {
                        "description": "查询请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新 AlerTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerTemplate 管理"
                ],
                "summary": "更新 AlerTemplate",
                "parameters": [
                    {
                        "description": "更新请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertTemplateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除 AlerTemplate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerTemplate 管理"
                ],
                "summary": "删除 AlerTemplate",
                "parameters": [
                    {
                        "description": "删除请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/alertTemplate/:id/copy": {
            "post": {
                "description": "拷贝一个已有的 AlerTemplate，需指定新名称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerTemplate 管理"
                ],
                "summary": "拷贝 AlerTemplate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "被拷贝模板的 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "拷贝请求参数（name 为新模板名称）",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertTemplateCopyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "拷贝成功，返回新创建的模板",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Matcher": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "标签名",
                    "type": "string"
                },
                "type": {
                    "description": "操作符: =, !=, =~, !~",
                    "type": "string"
                },
                "value": {
                    "description": "标签值",
                    "type": "string"
                }
            }
        },
...
        "types.AlertSilenceCreateRequest": {
            "type": "object",
            "required": [
                "comment",
                "endsAt",
                "startsAt",
                "status",
                "type"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 255
                },
                "endsAt": {
                    "type": "integer"
                },
                "fingerprint": {
                    "type": "string"
                },
                "matchers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Matcher"
                    }
                },
                "startsAt": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "type": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "types.AlertTemplateCopyRequest": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.AlertTemplateCreateRequest": {
            "type": "object",
            "required": [
                "name",
                "receiveId",
                "receiveIdType",
                "template"
            ],
            "properties": {
                "aggregationTemplate": {
                    "type": "string"
                },
                "alertChannelID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "receiveId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "receiveIdType": {
                    "type": "string",
                    "enum": [
                        "open_id",
                        "user_id",
                        "email",
                        "chat_id"
                    ]
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "types.AlertTemplateUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "receiveId",
                "receiveIdType",
                "template"
            ],
            "properties": {
                "aggregationTemplate": {
                    "type": "string"
                },
                "alertChannelID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "receiveId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "receiveIdType": {
                    "type": "string",
                    "enum": [
                        "open_id",
                        "user_id",
                        "email",
                        "chat_id"
                    ]
                },
                "template": {
                    "type": "string"
                }
            }
        },
...
        "types.IDRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "types.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}